+++ CHANGELOG.md
@@ -4,12 +4,14 @@
 ### notice
 - Amazon S3 inventory support is beta, requires diskover-web >= v1.5.0-rc9
 - --s3 requires index named diskover-s3-indexname
+- changes to diskover.cfg.sample, please copy over to your diskover.cfg and adjust for your env
 ### added
 - Amazon S3 inventory support - you can now import Amazon S3 inventory (CSV gzip format) to diskover ES index using --s3 cli arg and supplying 1 or multiple gzipped csv inventory files (see wiki or -h)
 - treewalk section with procs option to diskover.cfg.sample for setting number of parallel processes to tree walk (copy to diskover.cfg)
 - faster directory size calculations at end of crawl by reducing es update calls and using bulk update
 - maxsize in checkdupes section in diskover.cfg.sample - used for setting max file size to check for dupes (copy to diskover.cfg)
 - checkbytes in checkdupes section in diskover.cfg.sample - used for setting bytes to check at start and end of file before doing md5 sum check (copy to diskover.cfg)
+- new es optimization settings to elasticsearch section in diskover.cfg.sample - new settings for indexrefresh, disablereplicas, translogsize (copy from diskover.cfg.sample to your diskover.cfg)
 - autotagging to diskover_qumulo
 ### changed
 - switched to multiprocessing instead of threads for parallel tree walking